% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcoRt_fxns.R
\name{cort_repeat}
\alias{cort_repeat}
\title{calculate repatability of glucocorticoid response}
\usage{
cort_repeat(data = cort_sim2(sample_times = 6), boots = 100, perms = 0)
}
\arguments{
\item{data}{List object produced by 'cort_sim2()'.}

\item{boots}{Number of bootstraps for repeatability p-value estimation.}

\item{perms}{Number of permutations for repeatability p-value estimation.}
}
\value{
Returns a list that includes a single data frame with all repeatability estimates and p-values along with three plot
objects summarizing the dataset.
}
\description{
This function takes simulated data from 'cort_sim2' and calculates repeatability estimates for various glucocorticoid measures.
}
\examples{
cort_repeat(data = cort_sim2(sample_times = 10), boots = 200)
}
