% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcoRt_fxns.R
\name{cort_sim1}
\alias{cort_sim1}
\title{simulate true phenotypic values for acute glucocorticoid response}
\usage{
cort_sim1(
  n = 20,
  base_mu = 5,
  base_sd = 2,
  base_min = 0.5,
  slope_mu = 15,
  slope_sd = 2,
  fastpct_mu = 0.55,
  fastpct_sd = 0.08,
  speed_mu = 25,
  speed_sd = 5,
  speed_min = 10,
  max_mu = log(45),
  max_sd = 0.21,
  maxtime_mu = 10,
  maxtime_sd = 3,
  maxtime_min = 5,
  return_mu = 90,
  return_sd = 15,
  extra_n = 500,
  cor_base_speed = 0,
  cor_base_max = 0,
  cor_base_maxtime = 0,
  cor_base_return = 0,
  cor_base_slope = 0,
  cor_base_fastpct = 0,
  cor_speed_max = 0,
  cor_speed_maxtime = 0,
  cor_speed_return = 0,
  cor_speed_slope = 0,
  cor_speed_fastpct = 0,
  cor_max_maxtime = 0,
  cor_max_return = 0,
  cor_max_slope = 0,
  cor_max_fastpct = 0,
  cor_maxtime_return = 0,
  cor_maxtime_slope = 0,
  cor_maxtime_fastpct = 0,
  cor_return_slope = 0,
  cor_return_fastpct = 0,
  cor_slope_fastpct = 0
)
}
\arguments{
\item{n}{The number of individuals to simulate.}

\item{base_mu, base_sd}{The mean and standard deviation for baseline glucocorticoids.}

\item{base_min}{The minimum possible baseline glucocorticoid value.}

\item{slope_mu, slope_sd}{The mean and standard deviation for the time at which the animal reaches the
end of the initial 'fast increase' in glucocorticoids.}

\item{speed_mu, speed_sd}{The mean and standard deviation for the time to reach the maximum glucocorticoid level.}

\item{speed_min}{The minimum time required to reach the maximum glucocorticoid level.}

\item{max_mu, max_sd}{The mean and standard deviation for the total acute glucocorticoid increase (this is added to baseline).}

\item{maxtime_mu, maxtime_sd}{The mean and standard deviation for the time spent at the maximum glucocorticoid level.}

\item{maxtime_min}{The minimum amount of time spent at the maximum glucocorticoid level.}

\item{return_mu, return_sd}{The mean and standard deviation for the time required to return to baseline after maximum.}

\item{extra_n}{The number of extra simulations saved. These are used to add noise in 'cort_sim2'.}

\item{cor_x_y}{The population level correlation between each pair of parameters described above;
these work best by changing only a couple at a time or else it becomes impossible to estimate multivariate normal distributions.}
}
\value{
A list with three data frames for i) the simulated data, ii) the variance-covariance matrix, iii) the 'extra' simulations.
}
\description{
This function simulates glucocoriticoid response parameters for a population of animals of size n.
These can be thought of as the 'true' phenotypes of these individuals. A total of seven parameters
are sampled and together these parameters can be used to fit the shape of the glucocorticoid response
for each animal. The mean and standard deviation of each parameter along with the covariance between
each pair of parameters can be adjusted. A set of 'extra' animals are also simulated for use with the
subsequent function that simulates expressed glucocorticoid responses.
}
\examples{
cort_sim1(n = 50, base_mu = 10, speed_mu = 35, cor_base_speed = 0.5)
}
