% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcoRt_fxns.R
\name{cort_sim2}
\alias{cort_sim2}
\title{simulate observed acute glucocorticoid responses}
\usage{
cort_sim2(
  data = cort_sim1(),
  bleed_times = c(1, 15, 30),
  base_error = 0.8,
  speed_error = 0.6,
  max_error = 0.5,
  maxtime_error = 0.5,
  return_error = 0.5,
  slope_error = 0.5,
  fastpct_error = 0.5,
  sample_times = 2,
  assay_error = 0.2,
  timecourse_max = 170,
  performance_contributions = c(0, 0, 0, 0, 0, 1, 0, 1),
  sm_span = 0.5
)
}
\arguments{
\item{data}{List object created by 'cort_sim1()' or by default calls 'cort_sim1()' with default values.}

\item{bleed_times}{Numeric vector indicating times that blood samples would be collected at; used to save downsampled dataset.}

\item{sample_times}{Number of times that each animal is sampled.}

\item{assay_error}{The amount of measurement error as a percentage of the true value.}

\item{timecourse_max}{Number of timesteps (e.g. minutes) to simulate data for.}

\item{performance_contributions}{Numeric vector with 8 values indicating the relative contributions to fitness/performance from
base, speed, max, maxtime, return, slope, fastpct, and random error.}

\item{sm_span}{Smoothing parameter for loess regression. Smaller values are more wiggly}

\item{x_error}{Specify the amount of variation in each observed parameter that results from a random draw from the population. The remaining
amount of variation (1 - x_error) results from the 'true' phenotypic value of this individual. Extra simulations from 'cort_sim1()' are used
to add noise so that the population level covariance structure is maintained.}
}
\value{
A list with five dataframes storing the simulated data i) simulated_dataset_long: the downsampled dataset with measures taken with
error at each time point specified by bleed_times; ii) timecourse_long: the complete simulated dataset with measures taken
at every timestep; iii) rank_timecourse: the complete timecourse converted to a ranked order at each step; iv) AUC_measures: area under
the curve measures for each response calculated in a variety of ways for the full and downsampled data set; v) true_values: the true
phenotypic values used to simulate each animal joined to the fitness values.
}
\description{
This function starts with the phenotypic values for a population produced by 'cort_sim1' and simulates an arbitrary number
of actual expressions of acute glucocorticoid responses. Noise is added to each of the true parameters and the amount of noise can be
specified. This should be considered within-individual variation in the expression of underlying 'true' values. The sampled parameters
are used to fit smoothed response curves that allow measurement of glucocorticoids at each time step of the response. A second parameter can
also add noise resulting from measurement error, which is only added to a down-sampled data set with values saved at specified times
that would represent the actual time points that an empirical study might collect samples at.
}
\examples{
cort_sim2(data = cort_sim1(n = 50), bleed_times = c(2, 30, 60), sample_times = 5)
}
