% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcoRt_fxns.R
\name{plot_cort_sim}
\alias{plot_cort_sim}
\title{plot summary of glucocorticoid simulation}
\usage{
plot_cort_sim(data = cort_sim2())
}
\arguments{
\item{data}{Simulated data from 'cort_sim2'.}
}
\value{
Returns a single ggplot multi-panel plot.
}
\description{
This plots the output of a simulation from 'cort_sim1' as a three panel plot that shows the down-sampled data set.
}
\examples{
plot_cort_sim()
}
