% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcoRt_fxns.R
\name{plot_cort_sim_1}
\alias{plot_cort_sim_1}
\title{plot glucocorticoid response}
\usage{
plot_cort_sim_1(data = cort_sim2(), x_max = 60, y_max = 85)
}
\arguments{
\item{data}{Simulated data from 'cort_sim2'.}

\item{x_max}{Upper limit of x axis.}

\item{y_max}{Upper limit of y axis.}
}
\value{
Returns a single plot.
}
\description{
This plots a single panel showing the simulated glucocorticoid responses.
}
\examples{
plot_cort_sim1()
}
