% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcoRt_fxns.R
\docType{data}
\name{rwbb}
\alias{rwbb}
\title{Red winged blackbird corticosterone responses.}
\format{
A data frame with 408 rows and 12 variables:
\describe{
\item{id}{unique identity of this animal}
\item{time_sec}{time sampling starts, in seconds after midnight}
\item{sex}{sex of this individual}
\item{mass}{mass, in grams}
\item{tarsus}{tarsus length, in mm}
\item{wing}{flattened wing chord, in mm}
\item{h_bill}{head plus bill length, in mm}
\item{lh_stg}{life history stage, earl-breeding; late-breeding; or molt}
\item{age}{age, A = adult, SY = second year, U = unknown}
\item{lat_sec}{latency after capture for measurement, in seconds}
\item{cort}{corticosterone measurement, ng/ul}}
}
\usage{
rwbb
}
\description{
A dataset containing 7-timepoint stress series measures of red winged blackbirds.
}
\keyword{datasets}
