% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcoRt_fxns.R
\docType{package}
\name{simcoRt}
\alias{simcoRt}
\title{simcoRt: simulating physiological response data}
\description{
This package simulates response curve shapes for a population of animals and then uses those 'true'
values ot simulate observed responses with measurement error and within-individual variation. Subsequent
functions can calculate repeatability and make simple plots. The goal of the package is to simulate
different scenarios of variation in physiological responses (i.e., acute stress response) to allow for
examination of how sample design and sample sizes will allow differences to be detected.
}
\section{Author}{


Conor Taff cct663@gmail.com
}

\section{simcoRt functions}{
\preformatted{cort_sim1: simulates true phenotypic values for a population of animals with given parameters

cort_sim2: simulates arbitrary number of individual physiological responses for each animal with adjustible variation

cort_repeat: calculates repeatability of various common glucocorticoid measures from simulated datasets

plot_cort_sim: plotting wrapper function for multipanel summary plot

plot_cort_sim1: plotting wrapper function for single response curve plot
}
}

